Aprimorações na validação de id e pequenos detalhes nos deletes

+++ src/controllers/choice.contoller.js
@@ -49,6 +49,10 @@ const findChoice = async (req, res) => {
 
     const { id } = req.params
 
+    const validation = idSchema.validate({ id })
+
+    if (validation.error) return res.status(422).send(validation.error.details[0].message)
+
     try {
         const pollExists = await db.collection('polls').findOne({ _id: ObjectId(id) })
         if (!pollExists) return res.sendStatus(404)
@@ -64,7 +68,15 @@ const findChoice = async (req, res) => {
 
 const deleteChoice = async (req, res) => {
     const { id } = req.params
+
+    const validation = idSchema.validate({ id })
+
+    if (validation.error) return res.status(422).send(validation.error.details[0].message)
+
     try {
+        const choiceExists = await db.collection('choices').findOne({ _id: ObjectId(id) })
+        if (!choiceExists) return res.sendStatus(404)
+
         await db.collection('choices').deleteOne({ _id: ObjectId(id) })
     } catch (error) {
         console.log(error);
